+++ RELEASE_NOTES.md
@@ -24,10 +24,10 @@ wrapper. Use `compactMap` directly.
 - The entire `Onboarding` namespace has been removed; use `Tutti` instead.
 - `DelayOperator` has new delay functions with more time units. As a result, the
 original delay function is removed.
-- `CornerRadiusStyle` has been converted from a protocol to a struct. It is also
-renamed to `CornerRadius`.
+- `CornerRadiusStyle` has been renamed to `CornerRadius` and changed to a struct.
 - `UIView+CornerRadiusStyle` has been slightly refactored with signature changes.
 - `ParallaxEffectDefault` has been renamed to `StandardParallaxEffect`.
+- `ShadowStyle` has been renamed to `Shadow` and changed to a struct.